refactor to_date
